[FIX] stock_helper: fix tests based on modules installed
Since this change in v17
https://github.com/odoo/odoo/commit/439ca89a68fe957adbe0f7a6147047593155aa9f
_get_rule fallbacks to location's warehouse to check for routes.
This makes that the receipt route is found by _get_source_location_from_route.
However, this is the case only with `stock` module installed, because
as soon as `mrp` or `purchase` is installed that first step of the receipt
route get removed in favor of manufacture or purchase routes. See
https://github.com/odoo/odoo/commit/d25d320e0e870980971de945719f7b7d5deadd3c